+++ backend/webui_service/webui_init.go
@@ -32,7 +32,6 @@ import (
 	"github.com/omec-project/webconsole/configmodels"
 	"github.com/omec-project/webconsole/dbadapter"
 	gServ "github.com/omec-project/webconsole/proto/server"
-	"github.com/omec-project/webconsole/ui"
 	"github.com/sirupsen/logrus"
 	"github.com/urfave/cli"
 )
@@ -170,7 +169,7 @@ func (webui *WEBUI) Start() {
 
 	/* First HTTP Server running at port to receive Config from ROC */
 	subconfig_router := logger_util.NewGinWithLogrus(logger.GinLog)
-	ui.AddUiService(subconfig_router)
+	AddUiService(subconfig_router)
 	configapi.AddServiceSub(subconfig_router)
 	configapi.AddService(subconfig_router)
